+++ backend/src/test/java/turip/favorite/service/FavoriteServiceTest.java
@@ -1,4 +1,4 @@
-package turip.favorite;
+package turip.favorite.service;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -21,6 +21,7 @@ import turip.content.repository.ContentRepository;
 import turip.creator.domain.Creator;
 import turip.exception.BadRequestException;
 import turip.exception.NotFoundException;
+import turip.favorite.FavoriteService;
 import turip.favorite.controller.dto.request.FavoriteRequest;
 import turip.favorite.controller.dto.response.FavoriteResponse;
 import turip.favorite.domain.Favorite;
